+++ changelog.md
@@ -1,6 +1,11 @@
 # Changelog
 All notable changes to this project will be documented in this file. Major releases are documented [here](https://github.com/jgyates/genmon/releases)
 
+## V1.12.23 - 2019-03-06
+- Minor update to email password validation
+- Change to allow test email to be sent without a password 
+- Added option in Evolution Enhanced Exercise Add On to allow use of generator time instead of system time.
+
 ## V1.12.22 - 2019-03-02
 - Fix bug in H-Panel code that was introduced with V1.12.21